fix 后台分销显示 & 前台slaves

+++ models/flowhistory.js
@@ -10,7 +10,7 @@ module.exports = function(sequelize, DataTypes) {
     state: { type: DataTypes.INTEGER, allowNull: false },
     type: { type: DataTypes.STRING, allowNull: true },
     typeId: { type: DataTypes.INTEGER, allowNull: true },
-    amount: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 0 },
+    amount: { type: DataTypes.DECIMAL(10, 2), allowNull: false, defaultValue: 0.00 },
     comment: { type: DataTypes.STRING, allowNull: false },
     source: {
       type: DataTypes.VIRTUAL,
